added todos made engine methods private
