--- src/main/res/bruh.txt
@@ -1,1 +1,0 @@
-bruh
